+++ src/config/init/initApp.ts
@@ -6,7 +6,6 @@ import {Express} from 'express';
 const sleep = promisify(setTimeout);
 
 export const initApp = async (app: Express, mongoose: Mongoose, logger: ILogger) => {
-
     await sleep(20); // <= wait for app launch (db, cache...)
     logger.info('### App init Done.');
 };