+++ src/shared/tools.ts
@@ -1,5 +1,4 @@
 export class Tools {
-
     static GetHostname(): string {
         const os = require('node:os');
         return os.hostname();
@@ -9,5 +8,4 @@ export class Tools {
         const os = require('node:os');
         return os.cpus().length;
     }
-
 }